프로젝트 배포 준비

+++ src/main/java/BackendProject/domain/Quote.java
@@ -6,6 +6,6 @@ import lombok.Data;
 
 @Data
 public class Quote {
-    private Long qday;
+    private Integer qday;
     private String content;
 }
